fix: No need to "hide" missing files
Since we're now calling get_go_version_from_mod_file with paths we
found, we do not expect to have to handle missing files anymore.

If we want to handle missing files, we can do so in the function that
calls get_go_version_from_mod_file, not in the function itself.