feat: Add support for nested go.mod files (#102)
* refactor: Make the file a parameter of get_go_version_from_mod_file

This way this function can be tested more easily.

* refactor: Reorder functions

Move the helper functions that fetch the latest Go version and configure
logging to the top. This way the functions that do the actual magic are
closer together.

* refactor: Make the file a parameter of update_go_version_in_mod_file

This way this function can be tested more easily.

* feat: Find all go.mod files recursively

Closes issue #95.

* fix: No need to "hide" missing files

Since we're now calling get_go_version_from_mod_file with paths we
found, we do not expect to have to handle missing files anymore.

If we want to handle missing files, we can do so in the function that
calls get_go_version_from_mod_file, not in the function itself.