build: Replace requirements.txt with pyproject.toml
The immediate benefit is that we can make a destinction between the
requirements for the runtime environment and the requirements for the
development environment.

This is also reflected in the `actions.yml` file, where we can now use
`pip install .` to install all dependencies required to run the action.
In our tests (`python.yml`) we can use `pip install .[dev]` to also
install the requirements to develop (and test) the tool.

Note that this still does NOT tie us to a specific package manager as
pyproject.toml is widely supported.